requirements.txt
azure-functions
azure-communication-sms